# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nitip Core REST API server.",
        "title": "Nitip Core API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@nitip.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/admin/configs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all dynamic configuration values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Admin] System Config"
                ],
                "summary": "[ADMIN] List all configs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_domain_config.Config"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/configs/{key}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set or update a specific configuration key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Admin] System Config"
                ],
                "summary": "[ADMIN] Update a config value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Config Key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New config details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_config.updateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/kyc/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of KYC submissions waiting for review",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Admin] KYC Review"
                ],
                "summary": "[ADMIN] List pending KYC submissions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_domain_kyc.KycSubmission"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/kyc/{id}/review": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approve or reject a user's identity verification documents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Admin] KYC Review"
                ],
                "summary": "[ADMIN] Review a KYC submission",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "KYC ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review decision",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_kyc.ReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
...
        "/kyc/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the identity verification status for the logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] KYC"
                ],
                "summary": "Get current user KYC status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_domain_kyc.KycSubmission"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
        "/kyc/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register id card number and upload images for identity verification",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] KYC"
                ],
                "summary": "Submit KYC documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Card Number",
                        "name": "id_card_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "ID Card Image",
                        "name": "id_card",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Selfie Image",
                        "name": "selfie",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_domain_kyc.KycSubmission"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_codecoffy_nitip-core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_codecoffy_nitip-core_pkg_response.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_codecoffy_nitip-core_pkg_response.envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.ValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_domain_config.Config": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "internal_domain_config.updateConfigRequest": {
            "type": "object",
            "required": [
                "value"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "internal_domain_kyc.KycSubmission": {
            "type": "object",
            "properties": {
                "admin_note": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_card_image_url": {
                    "type": "string"
                },
                "id_card_number": {
                    "type": "string"
                },
                "selfie_image_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_domain_kyc.ReviewRequest": {
            "type": "object",
            "properties": {
                "approved": {
                    "type": "boolean"
                },
                "note": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}